Added a Thermostat for the shops electric heater.
It's working but about half of it is not being used. Still using some of Chris's old Dallas Temp Sensor stuff. Added a 2x16 display, pot and RTC. The display shows the time the Set Temp and the Shop Temp. The pot changes the Set Temp. At this time the RTC is only used to show the current time. Using the pot the Set Temp can go from 40° to 80°. The pots outer two legs go from VCC to GND and the center wiper pin goes to A0 on the Arduino. The math for that is A0 Reading*((40/1024)+40). No real PID here.