Shop_Electric_Heater_V1_12-27-16 updated version of SimpleThermostat
Made major changes to Thermostat program. Removed older Dallas Temp routines and added Heater status (on/of)f and Temp Setting  along with ShopTemp passed back to MySensors. Filename is now = to Sketch name passed to MySensors. Needs to be kept in sync so we know what Sketch is loaded on the system. Device ID = 26