More changes to V4-2.8 and V4.2.8.1
V4.2.8.1 is the Sketch currently running on the production Mega. Replaced the delay(1000) which was used to wait for both the temp sensor settle time and the Flow sensors "read pluses for 1 sec". Turns out the Flow sensor works on interrupts which the delay statement blocks. Also changed the interrupt # from 0 to 2. Pin 21 on the Mega is interrupt 2. On the pressure sensor I changed the formula for pressure from tankPumpPressure = ((voltage - 111) / 7); to tankPumpPressure = ((rawSensorValue - 111) / 7);. Voltage is a number from 0-5 volts whereas rawSensorValue is a number from 0-1023, big difference. Will look into using the map command insted of a bunch of math to convert 0-1023 to a range of 40° to 80° for a thermostat setting skipping the whole voltage thing.