Added Temp Sensor to Lower Solar Panel Inlet
Having problem with supply line from panel pump to lower solar panel inlet freezing up. Thought is there will be less chance of freeze up if we wait until the lower part of the solar is above 40° to turn on the pump. That's the thought. Hopefully fix minPumpOnTime 2 min clock.